shopgate update
git-svn-id: https://svn.modified-shop.org/svn/modified-shop/tags/modified-shop-1.06@9587 79ac0487-9261-4f26-a34d-858dee0dcce0

+++ inc/xtc_remove_order.inc.php
@@ -24,10 +24,12 @@
     xtc_db_query("DELETE FROM ".TABLE_ORDERS_PRODUCTS_DOWNLOAD." WHERE orders_id = '".(int)$order_id."'");
 
     /******** SHOPGATE **********/
-    $sql_select="SHOW TABLES LIKE '".TABLE_SHOPGATE_ORDERS."'";
-    $query = xtc_db_query($sql_select);
-    if(xtc_db_num_rows($query) > 0) {
-      xtc_db_query("DELETE FROM ".TABLE_SHOPGATE_ORDERS. " WHERE orders_id = '".(int)$order_id."'");
+    if(defined('MODULE_PAYMENT_SHOPGATE_STATUS') && MODULE_PAYMENT_SHOPGATE_STATUS=='True') {
+      $sql_select = "SHOW TABLES LIKE '" . TABLE_SHOPGATE_ORDERS . "'";
+      $query = xtc_db_query($sql_select);
+      if (xtc_db_num_rows($query) > 0) {
+        xtc_db_query("DELETE FROM " . TABLE_SHOPGATE_ORDERS . " WHERE orders_id = '" . (int)$order_id . "'");
+      }
     }
     /******** SHOPGATE **********/
   }
